Create tower_of_hanoi.py
 run python3 script.py.

Towers of Hanoi is an ancient mathematical puzzle that starts off with three stacks and many disks.

alt text

The objective of the game is to move the stack of disks from the leftmost stack to the rightmost stack.

The game follows three rules:

Only one disk can be moved at a time.
Each move consists of taking the upper disk from one of the stacks and placing it on top of another stack or on an empty rod.
No disk may be placed on top of a smaller disk.
In this project, we are going to use our knowledge of stacks to implement this game! Let’s get started!